+++ django/urls/__init__.py
@@ -3,6 +3,7 @@ from .base import (
     is_valid_path, resolve, reverse, reverse_lazy, set_script_prefix,
     set_urlconf, translate_url,
 )
+from .conf import include
 from .exceptions import NoReverseMatch, Resolver404
 from .resolvers import (
     LocaleRegexProvider, LocaleRegexURLResolver, RegexURLPattern,
@@ -15,6 +16,6 @@ __all__ = [
     'RegexURLPattern', 'RegexURLResolver', 'Resolver404', 'ResolverMatch',
     'clear_script_prefix', 'clear_url_caches', 'get_callable', 'get_mod_func',
     'get_ns_resolver', 'get_resolver', 'get_script_prefix', 'get_urlconf',
-    'is_valid_path', 'resolve', 'reverse', 'reverse_lazy', 'set_script_prefix',
-    'set_urlconf', 'translate_url',
+    'include', 'is_valid_path', 'resolve', 'reverse', 'reverse_lazy',
+    'set_script_prefix', 'set_urlconf', 'translate_url',
 ]
