+++ django/urls/conf.py
@@ -1,0 +1,50 @@
+"""Functions for use in URLsconfs."""
+from importlib import import_module
+
+from django.core.exceptions import ImproperlyConfigured
+
+from .resolvers import LocaleRegexURLResolver
+
+
+def include(arg, namespace=None):
+    app_name = None
+    if isinstance(arg, tuple):
+        # Callable returning a namespace hint.
+        try:
+            urlconf_module, app_name = arg
+        except ValueError:
+            if namespace:
+                raise ImproperlyConfigured(
+                    'Cannot override the namespace for a dynamic module that '
+                    'provides a namespace.'
+                )
+            raise ImproperlyConfigured(
+                'Passing a %d-tuple to include() is not supported. Pass a '
+                '2-tuple containing the list of patterns and app_name, and '
+                'provide the namespace argument to include() instead.' % len(arg)
+            )
+    else:
+        # No namespace hint - use manually provided namespace.
+        urlconf_module = arg
+
+    if isinstance(urlconf_module, str):
+        urlconf_module = import_module(urlconf_module)
+    patterns = getattr(urlconf_module, 'urlpatterns', urlconf_module)
+    app_name = getattr(urlconf_module, 'app_name', app_name)
+    if namespace and not app_name:
+        raise ImproperlyConfigured(
+            'Specifying a namespace in include() without providing an app_name '
+            'is not supported. Set the app_name attribute in the included '
+            'module, or pass a 2-tuple containing the list of patterns and '
+            'app_name instead.',
+        )
+    namespace = namespace or app_name
+    # Make sure the patterns can be iterated through (without this, some
+    # testcases will break).
+    if isinstance(patterns, (list, tuple)):
+        for url_pattern in patterns:
+            if isinstance(url_pattern, LocaleRegexURLResolver):
+                raise ImproperlyConfigured(
+                    'Using i18n_patterns in an included URLconf is not allowed.'
+                )
+    return (urlconf_module, app_name, namespace)
